requirements.txt
asgiref==3.4.1
black==21.12b0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.1
django-cors-headers==3.11.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.11
